made gif change happen first
will changing the gif first fix my problems I wonder.

It takes a while for the animated illustration to change in the popup--I think the images are just large files that take a hot second to load, but dunno if optimizing the images will actually fix the problem.

I'm stuck.