added 9-photo of dad
Adding assets/code for 9th response (photo of dad) and made text a little smaller.
NOTE to edit the js function making the popups. Add an extra paramater (boolean, shortPassage or somethin like that). On true, makes the text a lot bigger (for those responses that are like. Less than 5 words).
Would like to make something responsive with css but this is a temp fix.